Replace API Keys tab with SSH keys
